# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HAIROU BTP Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        clientName:
          type: string
        status:
          type: string
          enum: [PLANIFIE, EN_COURS, EN_PAUSE, TERMINE]
        budgetTotal:
          type: number
        budgetSpent:
          type: number
        progress:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        chefId:
          type: integer
        chefName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - budgetTotal
        - budgetSpent
        - progress
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/Task"
            expenses:
              type: array
              items:
                $ref: "#/components/schemas/Expense"
            workers:
              type: array
              items:
                $ref: "#/components/schemas/Personnel"
            messages:
              type: array
              items:
                $ref: "#/components/schemas/Message"
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        clientName:
          type: string
        status:
          type: string
          enum: [PLANIFIE, EN_COURS, EN_PAUSE, TERMINE]
        budgetTotal:
          type: number
        progress:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        chefId:
          type: integer
      required:
        - name
        - budgetTotal
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        clientName:
          type: string
        status:
          type: string
          enum: [PLANIFIE, EN_COURS, EN_PAUSE, TERMINE]
        budgetTotal:
          type: number
        progress:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        chefId:
          type: integer
    Task:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        title:
          type: string
        description:
          type: string
        assignedToId:
          type: integer
        assignedToName:
          type: string
        dueDate:
          type: string
          format: date
        priority:
          type: string
          enum: [BASSE, NORMALE, HAUTE, URGENTE]
        status:
          type: string
          enum: [A_FAIRE, EN_COURS, BLOQUEE, TERMINEE]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - priority
        - status
    CreateTaskRequest:
      type: object
      properties:
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
        assignedToId:
          type: integer
        dueDate:
          type: string
          format: date
        priority:
          type: string
          enum: [BASSE, NORMALE, HAUTE, URGENTE]
        status:
          type: string
          enum: [A_FAIRE, EN_COURS, BLOQUEE, TERMINEE]
      required:
        - projectId
        - title
        - priority
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assignedToId:
          type: integer
        dueDate:
          type: string
          format: date
        priority:
          type: string
          enum: [BASSE, NORMALE, HAUTE, URGENTE]
        status:
          type: string
          enum: [A_FAIRE, EN_COURS, BLOQUEE, TERMINEE]
...
    Personnel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        trade:
          type: string
        phone:
          type: string
        idNumber:
          type: string
        emergencyContact:
          type: string
        dailyWage:
          type: number
        contractType:
          type: string
          enum: [CDI, CDD, JOURNALIER]
        projectIds:
          type: array
          items:
            type: integer
        totalDaysWorked:
          type: integer
        totalPayOwed:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - trade
        - dailyWage
        - contractType
        - isActive
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        title:
          type: string
        category:
          type: string
          enum: [MATERIAUX, EQUIPEMENTS, TRANSPORT, CARBURANT, NOURRITURE, AUTRE]
        amount:
          type: number
        date:
          type: string
          format: date
        supplier:
          type: string
        receiptUrl:
          type: string
        addedById:
          type: integer
        addedByName:
          type: string
        status:
          type: string
          enum: [EN_ATTENTE, APPROUVEE, REJETEE]
        adminComment:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - category
        - amount
        - date
        - status
...
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        recipientId:
          type: integer
        recipientName:
          type: string
        projectId:
          type: integer
        projectName:
          type: string
        content:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - senderId
        - content
        - isRead
        - createdAt
...